app: fix int warnings in generic app files

+++ src/app-layer-parser.c
@@ -249,7 +249,7 @@ AppLayerParserThreadCtx *AppLayerParserThreadCtxAlloc(void)
     SCEnter();
 
     AppProto alproto = 0;
-    int flow_proto = 0;
+    uint8_t flow_proto = 0;
     AppLayerParserThreadCtx *tctx;
 
     tctx = SCMalloc(sizeof(*tctx));
@@ -275,7 +275,7 @@ void AppLayerParserThreadCtxFree(AppLayerParserThreadCtx *tctx)
     SCEnter();
 
     AppProto alproto = 0;
-    int flow_proto = 0;
+    uint8_t flow_proto = 0;
 
     for (flow_proto = 0; flow_proto < FLOW_PROTO_DEFAULT; flow_proto++) {
         for (alproto = 0; alproto < ALPROTO_MAX; alproto++) {
